+++ src/player-settings/index.js
@@ -1,0 +1,1 @@
+import "./pm-toggle-puzzle-piece-sound";