+++ src/player-settings/player-settings-defaults.ts
@@ -1,0 +1,5 @@
+const defaultSettings = {
+  playPuzzlePieceSound: false,
+};
+
+export default defaultSettings;